move `RelativeInterval` to a separate module
