+++ agent/provider/src/lib.rs
@@ -5,6 +5,7 @@ pub mod display;
 pub mod events;
 pub mod execution;
 pub mod hardware;
+mod interval;
 pub mod market;
 pub mod payments;
 pub mod provider_agent;
